move some functions to phoneNumbers.ts to follow the reference file
