+++ src/account/utils.ts
@@ -1,4 +1,4 @@
-import { parsePhoneNumber } from 'src/utils/parsePhoneNumber'
+import { parsePhoneNumber } from 'src/utils/phoneNumbers'
 
 const ADDRESS_LENGTH = 42 // TODO(ACT-1173): see if this can be replaced with a viem helper
